+++ src/test/java/game/OptionTest.java
@@ -1,9 +1,8 @@
-package client;
+package game;
 
 import static game.Option.findByStatus;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import game.Option;
 import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
